requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-multipart
faiss-cpu
sentence-transformers
pydantic
pdfplumber
python-dotenv
